+++ web-gui/src/app/app.module.ts
@@ -5,8 +5,8 @@ import {AppComponent} from './app.component';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {LayoutModule} from '@angular/cdk/layout';
 import {AppRoutingModule} from './app-routing.module';
-import {LoginComponent} from './login/login.component';
-import {MaterialComponentsModule} from "./material-components/material-components.module";
+import {LoginComponent} from './components/login/login.component';
+import {MaterialComponentsModule} from "./modules/material-components/material-components.module";
 import {FormsModule} from "@angular/forms";
 import {HttpClientModule} from "@angular/common/http";
 import {JwtHelperService, JwtModule} from "@auth0/angular-jwt";
