+++ web-gui/src/app/modules/admin/admin-dashboard/admin-dashboard.component.scss
