+++ web-gui/src/app/modules/admin/admin-nav/admin-nav.component.scss
