+++ web-gui/src/app/modules/admin/admin-nav/admin-nav.component.ts
@@ -3,7 +3,7 @@ import {BreakpointObserver, Breakpoints} from '@angular/cdk/layout';
 import {Observable} from 'rxjs';
 import {map} from 'rxjs/operators';
 import {MatSnackBar} from "@angular/material";
-import {AuthService} from "../../service/auth.service";
+import {AuthService} from "../../../service/auth.service";
 
 @Component({
   selector: 'app-admin-nav',
