+++ web-gui/src/app/modules/professor/professor.module.ts
