+++ web-gui/src/app/modules/student/course-table/course-table.component.ts
