+++ web-gui/src/app/modules/student/student-course/student-course.component.scss
