+++ web-gui/src/app/modules/student/student-list/student-list.component.scss
