+++ web-gui/src/app/modules/student/student-search/student-search.component.scss
