+++ web-gui/src/app/modules/student/student-start/student-start.component.scss
